# change imviz reference data coords from GWCS to FITS WCS