--- a/jdaviz/configs/imviz/plugins/parsers.py
+++ b/jdaviz/configs/imviz/plugins/parsers.py
@@ -1,5 +1,5 @@
 import os
-
+import warnings
 import asdf
 import numpy as np
 from astropy import units as u
@@ -10,7 +10,6 @@ from astropy.utils.data import cache_contents
 
 from glue.core.data import Component, Data
 from gwcs.wcs import WCS as GWCS
-from stdatamodels import asdf_in_fits
 
 from jdaviz.core.registries import data_parser_registry
 from jdaviz.core.events import SnackbarMessage
@@ -214,7 +213,7 @@ def get_image_data_iterator(app, file_obj, data_label, ext=None):
 
     # load Roman 2D datamodels:
     elif HAS_ROMAN_DATAMODELS and isinstance(file_obj, rdd.DataModel):
-        data_iter = _roman_2d_to_glue_data(file_obj, data_label, ext=ext)
+        data_iter = _roman_asdf_2d_to_glue_data(file_obj, data_label, ext=ext)
 
     # load ASDF files that may not validate as Roman datamodels:
     elif isinstance(file_obj, asdf.AsdfFile):
@@ -249,7 +248,7 @@ def _parse_image(app, file_obj, data_label, ext=None, parent=None):
             # NOTE: if extending this beyond GWCS, the mouseover logic
             # for outside_*_bounding_box should also be updated.
             data.coords._orig_bounding_box = data.coords.bounding_box
-            data.coords.bounding_box = None
+
         if not data.meta.get(_wcs_only_label, False):
             data_label = app.return_data_label(data_label, alt_name="image_data")
 
@@ -354,53 +353,38 @@ def _jwst_to_glue_data(file_obj, ext, data_label):
     yield data, new_data_label
 
 
-def _jwst2data(file_obj, ext, data_label):
-    comp_label = ext.upper()
-    new_data_label = f'{data_label}[{comp_label}]'
-    data = Data(label=new_data_label)
-    unit_attr = f'bunit_{ext}'
-
+def _try_gwcs_to_fits_sip(gwcs):
+    """
+    Try to convert this GWCS to FITS SIP. Some GWCS models
+    cannot be converted to FITS SIP. In that case, a warning
+    is raised and the GWCS is used, as is.
+    """
     try:
-        # This is very specific to JWST pipeline image output.
-        with asdf_in_fits.open(file_obj) as af:
-            dm = af.tree
-            dm_meta = af.tree["meta"]
-            data.meta.update(standardize_metadata(dm_meta))
+        result = WCS(gwcs.to_fits_sip())
+    except ValueError as err:
+        warnings.warn(
+            "The GWCS coordinates could not be simplified to "
+            "a SIP-based FITS WCS, the following error was "
+            f"raised: {err}",
+            UserWarning
+        )
+        result = gwcs
 
-            if unit_attr in dm_meta:
-                bunit = _validate_bunit(dm_meta[unit_attr], raise_error=False)
-            else:
-                bunit = ''
+    return result
 
-            # This is instance of gwcs.WCS, not astropy.wcs.WCS
-            if 'wcs' in dm_meta:
-                data.coords = dm_meta['wcs']
 
-            imdata = dm[ext]
-            component = Component.autotyped(imdata, units=bunit)
+def _jwst2data(file_obj, ext, data_label):
+    if ext == 'data':
+        ext = 'sci'
 
-            # Might have bad GWCS. If so, we exclude it.
-            try:
-                data.add_component(component=component, label=comp_label)
-            except Exception:  # pragma: no cover
-                data.coords = None
-                data.add_component(component=component, label=comp_label)
+    hdu = file_obj[ext]
 
-    # TODO: Do not need this when jwst.datamodels finally its own package.
-    # This might happen for grism image; fall back to FITS loader without WCS.
-    except Exception:
-        if ext == 'data':
-            ext = 'sci'
-        hdu = file_obj[ext]
-        return _hdu2data(hdu, data_label, file_obj, include_wcs=False)
-
-    return data, new_data_label
+    return _hdu2data(hdu, data_label, file_obj, include_wcs=False)
 
 
 # ---- Functions that handle input from Roman ASDF files -----
 
-def _roman_2d_to_glue_data(file_obj, data_label, ext=None):
-
+def _roman_asdf_2d_to_glue_data(file_obj, data_label, ext=None, try_gwcs_to_fits_sip=False):
     if ext == '*':
         # NOTE: Update as needed. Should cover all the image extensions available.
         ext_list = ('data', 'dq', 'err', 'var_poisson', 'var_rnoise')
@@ -412,7 +396,12 @@ def _roman_2d_to_glue_data(file_obj, data_label, ext=None):
         ext_list = (ext, )
 
     meta = standardize_roman_metadata(file_obj)
-    coords = getattr(getattr(file_obj, 'meta', {}), 'wcs', None)
+    gwcs = meta.get('wcs', None)
+
+    if try_gwcs_to_fits_sip and gwcs is not None:
+        coords = _try_gwcs_to_fits_sip(gwcs)
+    else:
+        coords = gwcs
 
     for cur_ext in ext_list:
         comp_label = cur_ext.upper()
@@ -430,39 +419,6 @@ def _roman_2d_to_glue_data(file_obj, data_label, ext=None):
             prep_data_layer_as_dq(data)
 
         yield data, new_data_label
-
-
-def _roman_asdf_2d_to_glue_data(file_obj, data_label, ext=None):
-    if ext == '*':
-        # NOTE: Update as needed. Should cover all the image extensions available.
-        ext_list = ('data', 'dq', 'err', 'var_poisson', 'var_rnoise')
-    elif ext is None:
-        ext_list = ('data', )
-    elif isinstance(ext, (list, tuple)):
-        ext_list = ext
-    else:
-        ext_list = (ext, )
-
-    roman = file_obj.tree.get('roman')
-    meta = roman.get('meta', {})
-    coords = meta.get('wcs', None)
-
-    for cur_ext in ext_list:
-        if cur_ext in roman:
-            comp_label = cur_ext.upper()
-            new_data_label = f'{data_label}[{comp_label}]'
-            data = Data(coords=coords, label=new_data_label)
-
-            # This could be a quantity or a ndarray:
-            ext_values = roman.get(cur_ext)
-            bunit = getattr(ext_values, 'unit', '')
-            component = Component(np.array(ext_values), units=bunit)
-            data.add_component(component=component, label=comp_label)
-            data.meta.update(standardize_metadata(dict(meta)))
-            if comp_label == 'DQ':
-                prep_data_layer_as_dq(data)
-
-            yield data, new_data_label
 
 
 # ---- Functions that handle input from non-JWST, non-Roman FITS files -----