--- a/jdaviz/core/freezable_state.py
+++ b/jdaviz/core/freezable_state.py
@@ -219,10 +219,17 @@ class FreezableBqplotImageViewerState(BqplotImageViewerState, FreezableState):
 
                 pixel_ids = layer.layer.pixel_component_ids
                 world_bottom_left = data.coords.pixel_to_world(0, 0)
-                world_top_right = data.coords.pixel_to_world(
-                    layer.layer.data.shape[pixel_ids[1].axis] - 1,
-                    layer.layer.data.shape[pixel_ids[0].axis] - 1
-                )
+
+                if getattr(data.coords, 'bounding_box', None) is not None:
+                    bounds = data.coords.pixel_bounds
+                    world_top_right = data.coords.pixel_to_world(
+                        bounds[0][1] - 1, bounds[1][1] - 1
+                    )
+                else:
+                    world_top_right = data.coords.pixel_to_world(
+                        layer.layer.data.shape[pixel_ids[1].axis] - 1,
+                        layer.layer.data.shape[pixel_ids[0].axis] - 1
+                    )
 
                 if return_as_world:
                     x_min = min(x_min, world_bottom_left.ra.value)