… CHANGES.rst
… docs/imviz/displayimages.rst
… jdaviz/configs/imviz/tests/test_astrowidgets_api.py
… jdaviz/configs/imviz/tests/test_delete_data.py
… jdaviz/configs/imviz/tests/test_parser.py
… jdaviz/configs/imviz/tests/utils.py
… jdaviz/utils.py