### Background

Since #2179, Imviz displays images in the sky plane without distortion by creating a glue reference data layer that contains only coordinates (no image). The WCS-only layer was has its coordinates defined by a GWCS object. I chose GWCS over FITS WCS because it seemed to me at the time that encoding reference frame rotation in FITS WCS was nontrivial. Since then, I noticed that longitudinal pole rotation is supported by low level WCS.

### Why swap GWCS for FITS WCS in reference data?

1. The API for creating low level WCS is very very stable, whereas GWCS is evolving to support Roman. This lowers the risk of imviz maintenance burden on our end.
2. Low level WCS computes transformations in C, which will be as fast or faster than GWCS.

When we think about "linking performance", what we're usually talking about is limiting the number of links and removing redundant ones. The floor on the speed of calculations through links is set by the performance of the WCS transformations.

#### Profiling

I ran [profiling tests in this notebook](https://gist.github.com/bmorris3/3b63f6537b8649f2e50bd0b9edc25c3e) to do transformations using the coordinates from a simulated Roman WFI exposure from Build 16 (requires `roman_datamodels==0.22.0`). I compared the GWCS performance with FITS SIP computed by calling the `gwcs.to_fits_sip` method.

FITS WCS transformations are 1.5x faster than GWCS in pixel to world transformations computed on large grids (thousands of pixels per side). FITS WCS and GWCS have roughly equivalent performance in world to pixel transformations on large grids. When computing transformations on smaller numbers of pixels (10s-100s), FITS WCS outperforms GWCS by ~10x in both pixel to world and world to pixel transformations. The gap between FITS WCS and GWCS gets larger for smaller arrays due to overheads in GWCS, which internally initializes new objects on every transformation, so performance sees little improvement once arrays are smaller than ~100x100.

Aside: this profiling was performed on Roman observation coordinates with all of their complexity in both GWCS/FITS SIP representations. The reference data layer doesn't need any of that complexity – it's just the sky plane! The rotated WCS created for reference data in this PR is just RA---TAN/DEC--TAN ± rotation, without distortion.

![gwcs_fits_sip](https://github.com/user-attachments/assets/c4d56e03-a058-401c-bac0-26c1a1345412)

### Next steps

This PR swaps the reference data layer's WCS implementation from GWCS to FITS WCS. A second performance improvement will come from optionally approximating GWCS objects with FITS SIP when possible. When the SIP approximation can be made, this brings significant performance improvements for a small loss in accuracy.

The lower row of the profiling plot above shows the mean, standard deviation, and maximum of the difference between the GWCS and FITS SIP coordinates. The maximum difference is less about 0.1 pixels, where most FITS SIP coordinates are within 0.02 pixels of the GWCS coordinates. For reference, the astrometric precision of Roman source catalog entries from PSF fitting is expected to be about the same size (https://github.com/spacetelescope/romancal/pull/794).

Our goal is to let the user choose via API or UI whether or not to use this approximation for data that they've loaded, which is not enabled by this PR.

As part of this PR, I've added a hidden method `_try_gwcs_to_fits_sip` to the imviz parser which tries to approximate with FITS SIP for data that enters the parser with GWCS. This function never gets called in the code in this PR, but we can use it in the (near!) future when we add support for optional FITS SIP approx from GWCS input.


### Other notes

This PR removes a line in the imviz parser which removes the bounding box from GWCS on load. Among other changes, a similar change was suggested in #2887. That PR is not superseded by this one.

### Change log entry

- [ ] Is a change log needed? If yes, is it added to `CHANGES.rst`? If you want to avoid merge conflicts,
  list the proposed change log here for review and add to `CHANGES.rst` before merge. If no, maintainer
  should add a `no-changelog-entry-needed` label.

### Checklist for package maintainer(s)
<!-- This section is to be filled by package maintainer(s) who will
review this pull request. -->

This checklist is meant to remind the package maintainer(s) who will review this pull request of some common things to look for. This list is not exhaustive.

- [ ] Are two approvals required? Branch protection rule does not check for the second approval. If a second approval is not necessary, please apply the `trivial` label.
- [ ] Do the proposed changes actually accomplish desired goals? Also manually run the affected example notebooks, if necessary.
- [ ] Do the proposed changes follow the [STScI Style Guides](https://github.com/spacetelescope/style-guides)?
- [ ] Are tests added/updated as required? If so, do they follow the [STScI Style Guides](https://github.com/spacetelescope/style-guides)?
- [ ] Are docs added/updated as required? If so, do they follow the [STScI Style Guides](https://github.com/spacetelescope/style-guides)?
- [ ] Did the CI pass? If not, are the failures related?
- [ ] Is a milestone set? Set this to bugfix milestone if this is a bug fix and needs to be released ASAP; otherwise, set this to the next major release milestone. Bugfix milestone also needs an accompanying backport label.
- [ ] After merge, any internal documentations need updating (e.g., JIRA, Innerspace)?