@pllim 2025-03-05
In general I like the idea of using a simpler FITS WCS as the underlying "fake image" layer for Orientation. I am still curious about benchmark for (a) JWST, and (b) distorted HST/ACS exposures. These are accessible via Imviz example notebooks in this repo. Thanks!