@camipacifici 2025-03-18
I was testing this to see if it solves the problem with the viewer that sometimes freezes and instead found that the align by WCS is broken for the dataset in this notebook: https://github.com/spacetelescope/jdat_notebooks/blob/main/notebooks/cross_instrument/photometry_extragalactic/imviz.ipynb
Can you take a look please? I see a traceback when asking to link by WCS from the API.