@haticekaratay 2025-03-26
Thanks @bmorris3  for the detailed and clearly explained PR.
To test the performance difference, I used the Roman WFI calibration file (r0000101001001001001_0001_wfi01_cal.asdf). I compared the WCS from the ASDF file (gwcs.wcs._wcs.WCS) with a hand-built FITS WCS(similar to what’s being used in this PR)(astropy.wcs.WCS).
I ran both `pixel_to_world` and `world_to_pixel` for a 500×500 grid, repeating the tests 10 times and averaging the results.

Here are the results:
| Branch   | WCS Type             | `pixel_to_world` | `world_to_pixel` |
|----------|----------------------|------------------|------------------|
| This PR  | `astropy.wcs.WCS`    |    0.0389 s    |    0.0269 s  |
| Main     | `gwcs.wcs._wcs.WCS`  | 0.0480 s       | 0.0462 s      |

The FITS WCS used in the PR was consistently faster in both direction.