@pllim 2025-03-26
@haticekaratay , just curious, on that same machine and env, how is the timing for 10k x 10k image?

https://github.com/spacetelescope/jdaviz/pull/3483#issuecomment-2755095408