@haticekaratay 2025-03-26
> just curious, on that same machine and env, how is the timing for 10k x 10k image?

For 10k x 10k
WCS Type: <class 'astropy.wcs.wcs.WCS'>
pixel_to_world: mean = 15.2658s
world_to_pixel: mean = 15.2889s

WCS Type: <class 'gwcs.wcs._wcs.WCS'>
pixel_to_world: mean = 6.7963s
world_to_pixel: mean = 49.1957s