@pllim 2025-03-26
Thanks, @haticekaratay for https://github.com/spacetelescope/jdaviz/pull/3483#issuecomment-2755191090 .

p.s. The asymmetric performance from GWCS is kinda surprising.