@bmorris3 2025-03-28
@pllim RE https://github.com/spacetelescope/jdaviz/pull/3483#discussion_r2017507678: when WCS-linked, data with GWCS get called in both directions from pixel to world and world to pixel. If the world coordinates aren't defined, the pixel coordinates shouldn't be either.