@pllim 2025-03-28
Re: https://github.com/spacetelescope/jdaviz/pull/3483#issuecomment-2761334855 -- Strictly speaking that change is unrelated to the performance goals and is taking a useful debugging feature away, so I feel a little sad seeing it go.