@bmorris3 2025-03-28
@pllim and I iterated on Slack and discovered that we didn't understand each other. The test the the comment above is acting correctly, but the coord info behavior regresses for single, unlinked images in Imviz. This is fixed in 7ee3255.