@bmorris3 2025-04-07
@pllim – I agree with your comment above, "Pixel" shouldn't show in that row in that case. I've adjusted the code accordingly and addressed your other comments.