@pllim 2025-04-07
Thank you. The question about behavior change wrt `concepts/imviz_dithered_gwcs.ipynb` remains. I am not sure how Indigo wants to respond to that. Another dev will take over this review starting this coming sprint. Thank you for your patience.