@bmorris3 2025-04-10
@rosteen and I discussed the plan forward for coord info on mouseover at locations outside of the bounds of the selected/top image layer. We'll show world coords from the _reference data_ layer. This will be done in a follow-up PR as soon as this PR gets merged.