@rosteen 2025-04-10
I need to check to see if the last failure in https://github.com/spacetelescope/jdaviz/actions/runs/14383447257/job/40332784755?pr=3483 is known, I think the rest are all unrelated and known.